Cycles lights are pre-normalized, so this patch will tweak the lights so that they can be un-normalized. The following lights have been checked: - Spot - Sphere - Distant - Disk - Area
It also has been cross-referenced with the following renderers:
- Karma (Houdini 18.5)
- Renderman 23.5
- Arnold 6.2

This will cause a change in the looks for current light setups, be warned! Lights will need to be ticked with normalize on to have the previous look (some lights have been tweaked though exposure-wise to match the other renderers).

I believe I can see the disk-like characteristics of point/spot/sphere lights, compared to the other renderers the shape in the reflection doesn't align perfectly, and there is a little bit of energy loss. Stefan has a patch here with proper sphere calculation:
https://developer.blender.org/D4676